+++ core/model/src/commonMain/kotlin/net/maiatoday/tagspotter/core/model/Spot.kt
@@ -37,7 +37,7 @@ data class SpotImage(
     val timestamp: Long,
     val thumbnailPath: String = "",
     val isMain: Boolean = false,
-    val rating: Int = 0,
+    val rating: Long = 0L,
     val uuid: String = generateUuid(),
     val lastEditedAt: Long = timestamp,
     val ownerUid: String? = null
